libhwc1: Fix severe mmap() and munmap() length mismatches
The size of struct ExynosVideoMeta, used to determine lengths of
mappings in mmap() calls, is very large compared to the length used in
corresponding munmap() calls - 64 bytes. This causes massive memory
leakage.

Under the older BSP sources intended for Android 10, from where these
HWC1 sources were imported here, ExynosVideoMeta had a size of 400
bytes. Due to large additions to ExynosVideoMeta in these 13 sources,
it has grown to 0x2438, or 9272(!), bytes, causing 9272-64=9208 bytes
to be leaked every time mmap() and munmap() are called.

Because the leaking is so much more severe, it isn't so long in usage
with video playback before no more memory can be mapped, and that then
causes random soft reboots.

Fix this massive problem by matching lengths used for the mmap() and
munmap() calls.

Also, don't map the entire size of the ExynosVideoMeta struct - Since
only the 'eType' member in it is checked in between the mmap() and
munmap() calls, and it's right at the start of the struct, the size of
only the 'eType' member can be mapped (and unmapped), which is only a
tiny 4 bytes and seems to work fine.

For some trivia, this all looks like an oversight from someone at
Samsung. In some really old BSP sources [1], mmap(), as well as
munmap(), is called with a length of 64. When the mmap() calls were
changed for the lengths to be based on the size of ExynosVideoMeta, the
corresponding munmap() calls were missed and left at 64.

 1: https://github.com/LineageOS/android_hardware_samsung_slsi_exynos/blob/fb9857c5aa8923cc043b8f8a91bd24339c4bf528/libvppdisplay/ExynosDisplay.cpp#L1307

Change-Id: Ib3591a8e952e0e1dbf1ccfc9aec36d908ca1db29